Non-integer frame rates in Android HW encoder
Bug: webrtc:10812

+++ sdk/android/src/java/org/webrtc/BitrateAdjuster.java
@@ -15,7 +15,7 @@ interface BitrateAdjuster {
   /**
    * Sets the target bitrate in bits per second and framerate in frames per second.
    */
-  void setTargets(int targetBitrateBps, int targetFramerateFps);
+  void setTargets(int targetBitrateBps, double targetFramerateFps);
 
   /**
    * Should be used to report the size of an encoded frame to the bitrate adjuster. Use
